+++ hello/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
 from django.core.validators import MinValueValidator, MaxValueValidator
-from django.forms import ModelForm
 
 # Create your models here.
 class Greeting(models.Model):
@@ -25,13 +24,3 @@ class Review(models.Model):
     publish_date = models.DateField("Published on", auto_now_add=True, null=True)
     def __str__(self):
         return str(self.id)  + " " + self.headline
-
-class PostForm(ModelForm):
-    class Meta:
-        model = Post
-        fields = ['headline', 'price', 'content']
-
-class ReviewForm(ModelForm):
-    class Meta:
-        model = Review
-        fields = ['headline', 'review_text', 'score']
